Create Readme
This folder contains the coding work done while completing research at the University of Colorado, Denver in the Master of Integrated Sciences program. I am currently working to turn this into a pipeline in Matlab and Python that will automate sparse grid surrogate model construction with the intention of performing forward uncertainty quantification.